Refactor topK and relation function

+++ pkg/models/event.go
@@ -13,8 +13,8 @@ type WikiEditInfo struct {
 	Title       string `json:"TITLE"`
 	Url         string `json:"URL"`
 	EditsCount  int    `json:"EDITS_COUNT"`
-	WindowStart int    `json:"WINDOW_START"`
-	WindowEnd   int    `json:"WINDOW_END"`
+	WindowStart int64  `json:"WINDOW_START"`
+	WindowEnd   int64  `json:"WINDOW_END"`
 }
 
 type WikiEditInfoKey struct {
